Move namespace for API.
