+++ sdk/docs/SaasFulfillmentApi.md
@@ -30,7 +30,7 @@ Use this call to activate a subscription.
 //import com.azure.marketplace.ApiException;
 //import com.azure.marketplace.Configuration;
 //import com.azure.marketplace.auth.*;
-//import com.azure.marketplace.generated.SaasFulfillmentApi;
+//import com.azure.marketplace.api.SaasFulfillmentApi;
 
 ApiClient defaultClient = Configuration.getDefaultApiClient();
 
@@ -92,7 +92,7 @@ Unsubscribe and delete the specified subscription.
 //import com.azure.marketplace.ApiException;
 //import com.azure.marketplace.Configuration;
 //import com.azure.marketplace.auth.*;
-//import com.azure.marketplace.generated.SaasFulfillmentApi;
+//import com.azure.marketplace.api.SaasFulfillmentApi;
 
 ApiClient defaultClient = Configuration.getDefaultApiClient();
 
@@ -152,7 +152,7 @@ Enables the publisher to track the status of the specified triggered async opera
 //import com.azure.marketplace.ApiException;
 //import com.azure.marketplace.Configuration;
 //import com.azure.marketplace.auth.*;
-//import com.azure.marketplace.generated.SaasFulfillmentApi;
+//import com.azure.marketplace.api.SaasFulfillmentApi;
 
 ApiClient defaultClient = Configuration.getDefaultApiClient();
 
@@ -215,7 +215,7 @@ Gets the specified SaaS subscription. Use this call to get license information a
 //import com.azure.marketplace.ApiException;
 //import com.azure.marketplace.Configuration;
 //import com.azure.marketplace.auth.*;
-//import com.azure.marketplace.generated.SaasFulfillmentApi;
+//import com.azure.marketplace.api.SaasFulfillmentApi;
 
 ApiClient defaultClient = Configuration.getDefaultApiClient();
 
@@ -276,7 +276,7 @@ Use this call to find out if there are any private or public offers for the curr
 //import com.azure.marketplace.ApiException;
 //import com.azure.marketplace.Configuration;
 //import com.azure.marketplace.auth.*;
-//import com.azure.marketplace.generated.SaasFulfillmentApi;
+//import com.azure.marketplace.api.SaasFulfillmentApi;
 
 ApiClient defaultClient = Configuration.getDefaultApiClient();
 
@@ -337,7 +337,7 @@ Lists the outstanding operations for the current publisher.
 //import com.azure.marketplace.ApiException;
 //import com.azure.marketplace.Configuration;
 //import com.azure.marketplace.auth.*;
-//import com.azure.marketplace.generated.SaasFulfillmentApi;
+//import com.azure.marketplace.api.SaasFulfillmentApi;
 
 ApiClient defaultClient = Configuration.getDefaultApiClient();
 
@@ -398,7 +398,7 @@ Lists all the SaaS subscriptions for a publisher.
 //import com.azure.marketplace.ApiException;
 //import com.azure.marketplace.Configuration;
 //import com.azure.marketplace.auth.*;
-//import com.azure.marketplace.generated.SaasFulfillmentApi;
+//import com.azure.marketplace.api.SaasFulfillmentApi;
 
 ApiClient defaultClient = Configuration.getDefaultApiClient();
 
@@ -459,7 +459,7 @@ Use this call to update the plan, the user count (quantity), or both.
 //import com.azure.marketplace.ApiException;
 //import com.azure.marketplace.Configuration;
 //import com.azure.marketplace.auth.*;
-//import com.azure.marketplace.generated.SaasFulfillmentApi;
+//import com.azure.marketplace.api.SaasFulfillmentApi;
 
 ApiClient defaultClient = Configuration.getDefaultApiClient();
 
@@ -521,7 +521,7 @@ The resolve endpoint enables the publisher to resolve a marketplace token to a p
 //import com.azure.marketplace.ApiException;
 //import com.azure.marketplace.Configuration;
 //import com.azure.marketplace.auth.*;
-//import com.azure.marketplace.generated.SaasFulfillmentApi;
+//import com.azure.marketplace.api.SaasFulfillmentApi;
 
 ApiClient defaultClient = Configuration.getDefaultApiClient();
 
@@ -582,7 +582,7 @@ Update the status of an operation to indicate success or failure with the provid
 //import com.azure.marketplace.ApiException;
 //import com.azure.marketplace.Configuration;
 //import com.azure.marketplace.auth.*;
-//import com.azure.marketplace.generated.SaasFulfillmentApi;
+//import com.azure.marketplace.api.SaasFulfillmentApi;
 
 ApiClient defaultClient = Configuration.getDefaultApiClient();
