+++ sdk/src/main/java/com/azure/marketplace/api/SaasMarketplaceMeterApi.java
@@ -12,7 +12,7 @@
  * Licensed under the MIT license.
  */
 
-package com.azure.marketplace.generated;
+package com.azure.marketplace.api;
 
 import com.azure.marketplace.*;
 import com.google.gson.reflect.TypeToken;
